Checkout backend

See merge request lbaw/lbaw2324/lbaw23107!42

+++ app/Http/Controllers/api/CartProductController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
+use App\Http\Controllers\ProductController;
 use App\Models\Product;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
@@ -26,6 +27,9 @@ class CartProductController extends Controller
                 return response()->json(['error' => 'Product already in cart'], 400);
             } else {
                 $product = Product::find($productId);
+                if (ProductController::isProductSold($product)) {
+                    return response()->json(['error' => 'Product has been already sold'], 400);
+                }
                 if ($product == null) {
                     return response()->json(['error' => 'Product not found'], 404);
                 }
@@ -57,7 +61,9 @@ class CartProductController extends Controller
                 if ($product == null) {
                     return response()->json(['error' => 'Product not found'], 404);
                 }
-
+                if (ProductController::isProductSold($product)) {
+                    return response()->json(['error' => 'Product has been already sold'], 400);
+                }
                 $request->user()->cart()->detach($productId); // detach() is a method from BelongsToMany relationship (User.php
 
                 return response()->json(['success' => 'Product removed from cart successfully'], 200);
